Выносит ActionType и ActionCreator в отдельные файлы
