+++ src/index.jsx
@@ -3,7 +3,7 @@ import ReactDom from 'react-dom';
 import App from './components/app/app.jsx';
 import {createStore} from 'redux';
 import {Provider} from 'react-redux';
-import {reducer} from './reducer';
+import reducer from './redux/reducer';
 
 const store = createStore(
     reducer,
