+++ src/redux/action-type.js
@@ -1,0 +1,4 @@
+export const ActionType = {
+  INCREMENT_MISTAKES: `INCREMENT_MISTAKES`,
+  INCREMENT_STEP: `INCREMENT_STEP`
+};
