+++ src/redux/reducer.test.js
@@ -1,57 +1,7 @@
-import {reducer, ActionType, ActionCreator} from './reducer';
-
-const questions = [
-  {
-    type: `genre`,
-    genre: `rock`,
-    answers: [
-      {
-        id: 0,
-        src: `https://upload.wikimedia.org/wikipedia/commons/4/4e/BWV_543-fugue.ogg`,
-        genre: `rock`,
-      },
-      {
-        id: 1,
-        src: `https://ru-drivemusic.net/dl/U96HV5OOndnnyglkRXcuVQ/1581984519/download_music/novogodnie_pesni/abba-happy-new-year.mp3`,
-        genre: `blues`,
-      },
-      {
-        id: 2,
-        src: `https://mp3name.net/music/7694-twenty-one-pilots-stressed-out.mp3`,
-        genre: `jazz`,
-      },
-      {
-        id: 3,
-        src: `https://mp3name.net/music/41700-twenty-one-pilots-ride.mp3`,
-        genre: `rock`,
-      }
-    ],
-  },
-  {
-    type: `artist`,
-    song: {
-      artist: `Jim Beam`,
-      src: `https://upload.wikimedia.org/wikipedia/commons/4/4e/BWV_543-fugue.ogg`,
-    },
-    answers: [
-      {
-        id: 1,
-        picture: `//picsum.photos/134/134?r=34`,
-        artist: `Chivas Regal`,
-      },
-      {
-        id: 2,
-        picture: `//picsum.photos/134/134?r=35`,
-        artist: `Jack Daniels`,
-      },
-      {
-        id: 3,
-        picture: `//picsum.photos/134/134?r=36`,
-        artist: `Jim Beam`,
-      }
-    ],
-  }
-];
+import reducer from './reducer';
+import {ActionType} from './action-type';
+import {ActionCreator} from './action-creator';
+import {questionsForTest} from '../mocks/questions';
 
 describe(`Reducer work correctly`, () => {
   it(`Reducer without additional parameters should return initial state`, () => {
@@ -60,7 +10,7 @@ describe(`Reducer work correctly`, () => {
       maxMistakesCount: 3,
       questionId: 0,
       step: -1,
-      questions
+      questions: questionsForTest
     });
   });
 
@@ -68,27 +18,27 @@ describe(`Reducer work correctly`, () => {
     expect(reducer({
       step: -1,
       mistakesCount: 0,
-      questions,
+      questions: questionsForTest,
     }, {
       type: ActionType.INCREMENT_STEP,
       payload: 1,
     })).toEqual({
       step: 0,
       mistakesCount: 0,
-      questions,
+      questions: questionsForTest,
     });
 
     expect(reducer({
       step: -1,
       mistakesCount: 0,
-      questions,
+      questions: questionsForTest,
     }, {
       type: ActionType.INCREMENT_STEP,
       payload: 0,
     })).toEqual({
       step: -1,
       mistakesCount: 0,
-      questions,
+      questions: questionsForTest,
     });
   });
